[docs:docs/readme] - fixing readme with new information and adding some fix in api
